backend/requirements.txt
absl-py==0.8.0
amqp==2.5.1
astor==0.8.0
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-celery-beat==1.5.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
django-timezone-field==3.1
djangorestframework==3.10.3
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
importlib-metadata==0.23
imutils==0.5.3
ipython==7.8.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.13.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.6.5
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mlxtend==0.17.0
more-itertools==7.2.0
networkx==2.3
numpy==1.17.1
oauthlib==3.1.0
opencv-python==4.1.0.25
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.1
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-crontab==2.4.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.2.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
zipp==0.6.0
